refactor: 카테고리 리포지터리 패키지 위치 개선
