+++ src/test/java/com/yapp/artie/category/repository/CategoryRepositoryTest.java
@@ -3,6 +3,7 @@ package com.yapp.artie.category.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.yapp.artie.category.domain.Category;
+import com.yapp.artie.category.domain.CategoryRepository;
 import com.yapp.artie.gallery.domain.entity.artwork.Artwork;
 import com.yapp.artie.gallery.domain.entity.artwork.Tag;
 import com.yapp.artie.gallery.domain.entity.exhibition.Exhibition;
